pyproject.toml
[tool.poetry]
name = "pirank-recsys"
version = "0.1.0"
description = "PiRank-based recommendation system for document ranking"
authors = ["Александр Пичугин <your.email@example.com>"]
readme = "README.md"
packages = [{include = "pirank_recsys"}]

[tool.poetry.dependencies]
python = "^3.9"
torch = "^2.0.0"
pytorch-lightning = "^2.0.0"
hydra-core = "^1.3.0"
dvc = "^3.0.0"
mlflow = "^2.0.0"
onnx = "^1.14.0"
scikit-learn = "^1.3.0"
pandas = "^2.0.0"
numpy = "^1.24.0"
fire = "^0.5.0"
transformers = "^4.30.0"
datasets = "^2.12.0"
lightgbm = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.poetry.group.gpu.dependencies]
tensorrt = {version = "^8.6.0", optional = true}

[tool.poetry.extras]
gpu = ["tensorrt"]
